% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_all}
\alias{installed_all}
\title{Reinstalling installed packages after update}
\usage{
installed_all(path = NULL, updated = FALSE)
}
\arguments{
\item{path}{the path to where the packages are saved (stop at, but include, the "Versions" folder)}

\item{updated}{whether or not you have already updated R to a new version}
}
\description{
Looks up the packages that were installed on the previous version of R
}
