% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_cran}
\alias{installed_cran}
\title{Installed Packages-Versions Available on CRAN}
\usage{
installed_cran(path = NULL, updated = FALSE)
}
\arguments{
\item{path}{the path to where the packages are saved (stop at, but include, the "Versions" folder)}

\item{updated}{whether or not you have already updated R to a new version}
}
\description{
Looks up the package-version combinations available currently on CRAN
}
