% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_not_cran}
\alias{installed_not_cran}
\title{Installed Packages-Versions NOT Available on CRAN}
\usage{
installed_not_cran(path = NULL, updated = FALSE)
}
\arguments{
\item{path}{the path to where the packages are saved (stop at, but include, the "Versions" folder)}

\item{updated}{whether or not you have already updated R to a new version}
}
\description{
Looks up the packages-version not available on CRAN. In other words, these may be packages that were installed from GitHub
or another repository.
}
