% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_packages}
\alias{installed_packages}
\title{Looking Up Installed Packages}
\usage{
installed_packages(path = NULL, updated = FALSE)
}
\arguments{
\item{path}{the path to where the packages are saved (stop at, but include, the "Versions" folder)}

\item{updated}{whether or not you have already updated R to a new version}
}
\description{
Looks up the packages that were installed on machine
}
