% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_prior}
\alias{installed_prior}
\title{Save List of Packages Prior to Updating R}
\usage{
installed_prior(savefilepath, path = NULL, updated = FALSE)
}
\arguments{
\item{savefilepath}{the path where to save the list of packages on and not on CRAN}

\item{path}{the path to where the packages are saved (stop at, but include, the "Versions" folder)}

\item{updated}{whether or not you have already updated R to a new version}
}
\description{
Saves a list of packages that can be used after updating R
}
