% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed.R
\name{installed_versions}
\alias{installed_versions}
\title{Looks Up Installed Package Versions}
\usage{
installed_versions(path = NULL, updated = FALSE)
}
\description{
Looks up the package versions that are installed
}
